Add files via upload
This code work is the first attempt at using xgboost to predict which students will drop out of college. It is blind in that it attempts to predict based on the student being in semester for only two-weeks. A drop out is someone who does not attempt the next three semesters.